refactor options
