+++ examples/options/include/traccc/options/handle_argument_errors.hpp
