+++ examples/run/sycl/seq_example_sycl.sycl
@@ -25,8 +25,8 @@
 #include "traccc/efficiency/seeding_performance_writer.hpp"
 
 // options
-#include "traccc/full_tracking_input_options.hpp"
-#include "traccc/handle_argument_errors.hpp"
+#include "traccc/options/full_tracking_input_options.hpp"
+#include "traccc/options/handle_argument_errors.hpp"
 
 // Vecmem include(s)
 #include <vecmem/memory/host_memory_resource.hpp>
@@ -299,7 +299,7 @@ int main(int argc, char* argv[]) {
 
     // Add options
     desc.add_options()("help,h", "Give some help with the program's options");
-    traccc::add_full_tracking_input_options(desc);
+    traccc::full_tracking_input_config full_tracking_input_cfg(desc);
     desc.add_options()("run_cpu", po::value<bool>()->default_value(false),
                        "run cpu tracking as well");
 
@@ -307,7 +307,7 @@ int main(int argc, char* argv[]) {
     po::store(po::parse_command_line(argc, argv, desc), vm);
 
     // Read options
-    auto full_tracking_input_cfg = traccc::read_full_tracking_input_options(vm);
+    full_tracking_input_cfg.read(vm);
     auto run_cpu = vm["run_cpu"].as<bool>();
 
     // Check exception
